Ran flex struct and edited Structural_scaling.m
Ran the flex struct in HAWC and saved results. Also edited
Structural_scaling.m so that it creates a redesign_Blade_st.dat file
from the DTU_10MW_RWT_Blade_st.dat file and saves it in the correct
place. Also made it write a c2def file to the correct place.